robustify

+++ scripts/link-rstudio-snippets-and-bindings.sh
@@ -6,7 +6,9 @@ mkdir -p $HOME/.config/rstudio
 for item in snippets keybindings; do
     for snippet in rstudio/$item/*; do
         item_name="$(basename $snippet)"
-        ln -sf $HOME/.dotfiles/rstudio/$item/"$item_name"\
-            $HOME/.config/rstudio/$item/"$item_name"
+        if [[ -e $HOME/.dotfiles/rstudio/$item/"$item_name" ]];then
+            ln -sf $HOME/.dotfiles/rstudio/$item/"$item_name" \
+                $HOME/.config/rstudio/$item/"$item_name"
+        fi
     done
 done
